Move entries older than last year to changelog-old
The current changelog is *more than twice as long* as the "archive".
So it could really use some trimming. I propose to always keep the
current and previous years' entries.

First half of the change: Copy the old entries (2012 and older) over.
Once that's done, I'll truncate the current changelog file.